Fix sqlite autoincrement for orgs

+++ api/src/db/models/apps.py
@@ -1,7 +1,7 @@
 from datetime import datetime
-from src.db.models.__base__ import Base
+from sqlalchemy import DateTime, ForeignKey, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy import String, DateTime, BigInteger, ForeignKey, func
+from src.db.models.__base__ import Base
 
 
 class App(Base):
@@ -9,8 +9,8 @@ class App(Base):
 
     __tablename__ = 'apps'
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    id_org: Mapped[int] = mapped_column(BigInteger, ForeignKey('organizations.id'), nullable=False)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id_org: Mapped[int] = mapped_column(Integer, ForeignKey('organizations.id'), nullable=False)
     name: Mapped[str] = mapped_column(String(100), nullable=False)
 
     date_creation: Mapped[datetime] = mapped_column(DateTime, nullable=False, server_default=func.now())
